feat(neiji+app): 5分钟累计剂量实时上报基线
NeiJi 按原始 CPS 积分 30s/5min 窗，0x23(0x001E) 上传 D5(μSv×100)；
修复 once_cnt 三角累加与本底小数 CPS 截断；App 按设备时间入库并支持时差>5min 对时。

+++ probe/Neiji/NeiJi/Hardware/dose_rate/dose_rate.h
@@ -25,7 +25,7 @@ typedef struct {
 typedef struct {
     float current_avg_cps;
     float current_dose_rate;
-    int last_raw_cps;
+    float last_raw_cps;   /* 本底/死时间后有效 CPS（可含小数） */
     int boost_timer;
     int threshold_cps;
     int threshold_delta;
